api/requirements.txt
bert-serving-client==1.10.0
elasticsearch
Flask
flask-cors
deeppavlov
redis